requirements.txt
django
gunicorn
whitenoise
dj-database-url
psycopg2
django-ipware
requests
celery
redis
